requirements.txt
python-telegram-bot==20.3
openpyxl
Pillow
python-dotenv
